modernize project (#1)
- upgrade deps
- migrate to python 3.14
- implement adding torrent to the download via link
- remove ngrok/webhook support
- major refactoring
