--- .chglog/config.yml
@@ -1,29 +1,0 @@
-style: github
-template: CHANGELOG.tpl.md
-info:
-  title: CHANGELOG
-  repository_url: https://github.com/Vovcharaa/transmission-telegram-bot
-options:
-  commits:
-    filters:
-      Type:
-        - feat
-        - fix
-        - perf
-        - refactor
-  commit_groups:
-    title_maps:
-      feat: Features
-      fix: Bug Fixes
-      perf: Performance Improvements
-      refactor: Code Refactoring
-  header:
-    pattern: ^(\w*)(?:\(([\w\$\.\-\*\s]*)\))?\:\s(.*)$
-    pattern_maps:
-      - Type
-      - Scope
-      - Subject
-  notes:
-    keywords:
-      - BREAKING CHANGE
-      - BREAKING CHANGES
