+++ .env.example
@@ -1,0 +1,5 @@
+TELEGRAM_TOKEN=
+# default [{"name": "Default", "host": "localhost", "port": 9091}]
+TRANSMISSION_CLIENTS=[{"name": "Main", "host": "localhost", "port": 9091, "username": "username", "password": "qwerty123"}, {...}]
+# comma separated list of telegram user IDs
+WHITELIST=
