--- .github/workflows/docker-push.yml
@@ -1,36 +1,0 @@
-name: ci
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  docker-push:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v1.2.0
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v1.6.0
-        id: buildx
-        with:
-          install: true
-      - name: Login to GitHub Container Registry
-        uses: docker/login-action@v1.14.1
-        with:
-          registry: ghcr.io
-          username: ${{ github.repository_owner }}
-          password: ${{ secrets.CR_PAT }}
-      - name: Build and push
-        uses: docker/build-push-action@v2.9.0
-        with:
-          context: .
-          file: ./Dockerfile
-          platforms: linux/amd64,linux/arm64
-          push: true
-          pull: true
-          tags: |
-            ghcr.io/vovcharaa/transmission-telegram-bot:${{ github.sha }}
