+++ .github/workflows/release.yml
@@ -3,61 +3,32 @@ name: release
 on:
   push:
     tags:
-      - v*
+      - "*"
 
 jobs:
-  release:
+  deploy:
     runs-on: ubuntu-latest
-
     steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-      # create github release
-
-      - name: Get Tag Name
-        run: echo TAG=${GITHUB_REF##*/} >> $GITHUB_ENV
-
-      - name: Set up Go
-        uses: actions/setup-go@v2
-
+      - name: Checkout
+        uses: actions/checkout@v5
+      - name: Extract release version
+        run: echo "RELEASE_VERSION=${GITHUB_REF#refs/*/}" >> $GITHUB_ENV
       - name: Set up QEMU
-        uses: docker/setup-qemu-action@v1.2.0
-
+        uses: docker/setup-qemu-action@v3
       - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v1.6.0
-        id: buildx
-        with:
-          install: true
-
+        uses: docker/setup-buildx-action@v3
       - name: Login to GitHub Container Registry
-        uses: docker/login-action@v1.14.1
+        uses: docker/login-action@v3
         with:
           registry: ghcr.io
           username: ${{ github.repository_owner }}
-          password: ${{ secrets.CR_PAT }}
+          password: ${{ secrets.GITHUB_TOKEN }}
       - name: Build and push
-        uses: docker/build-push-action@v2.9.0
+        uses: docker/build-push-action@v6
         with:
           context: .
-          file: ./Dockerfile
-          platforms: linux/amd64,linux/arm64  # linux/386,linux/arm/v7
+          platforms: linux/amd64,linux/arm64
           push: true
-          pull: true
           tags: |
-            ghcr.io/vovcharaa/transmission-telegram-bot:latest
-            ghcr.io/vovcharaa/transmission-telegram-bot:${{ env.TAG }}
-
-      - name: Install git-chglog
-        run: |
-          wget -O git-chglog https://github.com/git-chglog/git-chglog/releases/download/0.9.1/git-chglog_linux_amd64
-          chmod +x git-chglog
-
-      - name: Generate Changlog
-        run: |
-          ./git-chglog "$TAG" | node -p 'require("fs").readFileSync(0).toString().trim()
-            .split("\n").slice(1).join("\n").trim()' > chglog.md
-      - name: Upload Github Release
-        run: gh release create "$TAG" -F chglog.md
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+            ghcr.io/aleksey925/transmission-telegram-bot:latest
+            ghcr.io/aleksey925/transmission-telegram-bot:${{ env.RELEASE_VERSION }}
