--- .vscode/settings.json
@@ -1,11 +1,0 @@
-{
-    "python.linting.pylintEnabled": false,
-    "python.linting.flake8Enabled": true,
-    "python.linting.enabled": true,
-    "python.autoComplete.extraPaths": [
-        "./transsmision-telegram-bot"
-    ],
-    "python.analysis.extraPaths": [
-        "./transsmision-telegram-bot"
-    ],
-}
