+++ Makefile
@@ -1,0 +1,5 @@
+img:
+	docker build -t transmission-telegram-bot:latest .
+
+lint:
+	pre-commit run --all-files
