--- bot.env
@@ -1,12 +1,0 @@
-TELEGRAM_TOKEN=
-# default [{"name": "Default", "host": "localhost", "port": 9091}]
-TRANSMISSION_CLIENTS=[{"name": "Main", "host": "localhost", "port": 9091, "username": "username", "password": "qwerty123"}, {...}]
-# comma separated list of chatIds
-WHITELIST=
-
-# How to receive updates from telegram
-# Allowed values: ngrok, webserver, polling
-# Default: polling
-UPDATER_TYPE=
-# If UPDATER_TYPE=webserver set this variable for autosetuping of webhook
-WEBHOOK_DOMAIN=
