+++ mise.toml
@@ -1,0 +1,10 @@
+[[env]]
+_.python.venv = { path = "{{ get_env(name='VENV_DIR_PATH', default='.venv') }}", create = true }
+
+[[env]]
+_.source = { path = ".env.sh", tools = true }
+
+[tools]
+pre-commit = "4.5.0"
+python = "3.14.0"
+uv = "0.9.9"
