--- requirements.txt
@@ -1,4 +1,0 @@
-pyngrok
-transmission-rpc
-python-telegram-bot
-python-dotenv
