--- setup.cfg
@@ -1,12 +1,0 @@
-[flake8]
-max-line-length = 88
-exclude =
-    .git,
-    __pycache__,
-    .pytest_cache,
-    venv
-extend-ignore =
-    # See https://github.com/PyCQA/pycodestyle/issues/373
-    E203,
-[tool.black]
-skip-string-normalization = true
