--- transmission-telegram-bot/__init__.py
@@ -1,3 +1,0 @@
-import logging
-
-logging.basicConfig()
