--- transmission-telegram-bot/__main__.py
@@ -1,4 +1,0 @@
-from .app import run
-
-if __name__ == "__main__":
-    run()
