+++ transmission_telegram_bot/__init__.py
@@ -1,0 +1,3 @@
+import logging
+
+logging.basicConfig()
