+++ transmission_telegram_bot/__main__.py
@@ -1,0 +1,4 @@
+from .app import run
+
+if __name__ == "__main__":
+    run()
